+++ game_core.py
@@ -1,48 +1,51 @@
 import sys
+
 import numpy as np
 
 
 class GameCore:
 
     def __init__(self):
-        self.input = []
-        self.input2 = []
-        self.i = 0
+        self.board_state = np.zeros(shape=(3, 3))
+        self.turn_counter = 0
 
-    def add_input(self):
-        while self.i < 9:
-            user_input = int(input())
-            if self.i % 2 == 0:
-                self.input.append(user_input)
-            elif self.i % 2 != 0:
-                self.input2.append(user_input)
-            if self.win_condition_check:
-                if self.i % 2 == 0:
-                    print("Wygral gracz numer 1")
-                else:
-                    print("Wygral gracz numer 2")
-                break
-            self.i += 1
-            print(self.input)
-            print(self.input2)
+    def _check_for_winning_conditions(self):
+        pass
+
+    def _is_occupied(self):
+        pass
+
+    def _is_input_valid(self, input):
+        if input >= 3:
+            return False
+        else:
+            return True
+
+    def _player_recognition(self):
+        if self.turn_counter % 2 == 0:
+            return 1
+        elif self.turn_counter % 2 != 0:
+            return 2
+        return None
+
+    def add_marker(self):
+
+        while self.turn_counter < 9:
+            response = (input())  # input with coordinates
+            row = int(response[0]) - 1
+            col = int(response[1]) - 1
+
+            if self._is_input_valid(row) and self._is_input_valid(col): # check whether coordinates don't exceed the grid size
+                player = self._player_recognition()  # which player's turn is it currently
+                self.board_state[row][col] = player
+                print(self.board_state)
+                self.turn_counter += 1
+            else:
+                print("Your coordinates exceed the grid size. Please change the input")
+
+
 
         sys.exit()
 
-    @property
-    def win_condition_check(self):
-        winning_combos = [
-            [1, 2, 3],
-            [1, 4, 7],
-            [1, 5, 9],
-            [2, 5, 8],
-            [4, 5, 6],
-            [3, 5, 7],
-            [7, 8, 9]
-        ]
-
-        if self.i >= 4:
-            for player_input in [self.input, self.input2]:
-                for combo in winning_combos:
-                    if all(x in player_input for x in combo):
-                        return True
-        return False
+    def _play_board_status(self):
+        pass
